project_1/pyproject.toml
[project]
name = "project-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.14",
    "langgraph>=1.1.8",
    "python-dotenv>=1.2.2",
]
